# Conflicts:
#	amplify/backend/api/restaurants/cli-inputs.json
#	amplify/backend/api/restaurants/schema.graphql
#	app/src/main/java/com/amplifyframework/datastore/generated/model/Order.java
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/cs389team4/needtofeed/utils/Utils.java
@@ -16,7 +16,7 @@ import androidx.appcompat.content.res.AppCompatResources;
 public class Utils {
 
     private Utils() {
-
+        // Empty private constructor to hide implicit public constructor
     }
 
     // Display toast message
@@ -25,7 +25,7 @@ public class Utils {
     }
 
     // Calculate delivery fee
-    public static double getDeliveryFee(double orderSubtotal) {
+    public static double getDeliveryFee(float orderSubtotal) {
         // Return 15 percent of order subtotal, at max of 5
         return Math.min(orderSubtotal * 0.15, 5.00);
     }
